pyproject.toml
[project]
name = "deadpool"
version = "0.1.0"
description = "Global Project repo for all Gen AI projects."
authors = [
    {name = "gshreyas8",email = "gotimukulshreyas@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "numpy (==1.26.4)",
    "tqdm (==4.66.6)",
    "ipykernel (==6.29.5)",
    "python-dotenv (==1.0.1)",
    "openai (>=1.58.1,<2.0.0)",
    "colorama (==0.4.6)",
    "rich (==13.9.4)",
    "lxml (==5.3.0)",
    "xmltodict (==0.14.2)",
    "langchain (==0.3.7)",
    "pandas (==1.5.3)",
    "scikit-learn (==1.2.2)",
    "fastapi (==0.115.6)",
    "langchain-openai (>=0.3.2,<0.4.0)",
    "langchain-community (==0.3.7)",
    "uvicorn (==0.34.0)",
    "httpx (==0.27.2)",
    "flask (>=3.1.0,<4.0.0)",
    "langchain-deepseek-official (>=0.1.0,<0.2.0)"
]

[tool.poetry]
package-mode = false 
[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pytest = "^7.0.0"
black = "^23.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
